Taking Long long int solve the problem when we take large value of A and B

+++ DSA/C_program/ArithmaticOperator.c
@@ -5,19 +5,19 @@
 
 #include<stdio.h>
 int main(){
-    int a , b  ;
+    long long int  a , b  ;
     printf("Enter a:");
-    scanf("%d",&a);
+    scanf("%lld",&a);
     printf("Enter b:");
-    scanf("%d",&b);
+    scanf("%lld",&b);
 
     int add = a + b;
     int sub = a - b;
-    int multi = a * b;
+    long long int multi = a * b;
     int quotiient = a/ b;
     int rem = a % b;
     //% is used for place holder ie why we use  %% in modulo
-    printf("add is %d \nsub is %d\nmulti is %d \ndivision is %d \nreminder ie a%%b %d \n",add , sub , multi , quotiient , rem);
+    printf("add is %d \nsub is %d\nmulti is %lld \ndivision is %d \nreminder ie a%%b %d \n",add , sub , multi , quotiient , rem);
 
 
 }
